Create basic analyzer library.
Prototype of a function that can load a plugin using ts-morph, typecheck it,
and extract plugin function signatures. This is incomplete, most notably
missing the ability to resolve external libraries (this appears to be a
bug in ts-morph's Deno support, not a fundamental limitation); however,
it should serve as a good start.