Disable concurrent invocations. (#11)
This broke logging, which works by capturing all calls made for the span
of a single invoke call. There's a deeper philosophical question to
answer about whether it's worth supporting concurrency within a worker at
all, compared to just creating multiple workers (which we likely want to
do anyway for resiliency). For the initial use cases though, we don't
expect it to matter, so just serialize.