Create invocation sandbox for controlling globals. (#13)
Restrict access to globals that allow messing with the worker, as well
as things we don't want to support initially like fetch.

For now, use the same restrictions for load and invoke, though we may
ultimately want to separate a LoadContext from InvocationContext.

Establish a pattern for, rather than a strict 'allow' or 'deny',
wrapping a global with a replacement function; implement this
initially for timers. Although there's not a strong use case for
setTimeout in real plugin code, it's very convenient in testing, and
several tests depend on it. This also serves as a demonstration for the
strategy for more complex cases, like adding headers in fetch.

Unfortunately, this still relies on invocations being serialized.
Ideally we could exit and reenter an invocation context on suspend and
resume -- this seems basically possible by overriding Promise and
utilizing  queueMicrotask, but this would need more testing. If we
could do this, we could solve things like log attribution by using the
global 'cid' symbol.