Create a pool of concurrent plugin workers. (#21)
First, add a 'concurrency' field to plugin configuration, which will
match an upcoming change to the resource format.

In PluginHost, maintain a pool of workers and dispatch invocations to
them one at a time. Handle load failures and aborted invocations (i.e.,
timeouts) by terminating the worker and restarting it.

The server type had the beginnings of a multi-host approach; this is no
longer needed, so remove it in favor of pooling at the worker level.

This trickiest part of this design is how to track the overall lifecycle
of the host to reflect potentially many workers. The choices here are
fairly arbitrary: a plugin is "loaded" when one worker is ready, and is
considered permanently failed if three loads fail in a row.

Finally, refactor InvocationContext; the previous enter/exit API
suggested (and was originally intended as) swapping in and out of
different call contexts as concurrent calls in the same worker suspended
and resumed. Hooking into every suspend and resume would be difficult,
and with the new worker pool we don't plan to do so. Therefore, make the
global environment API reflect that it's more of a one-shot
create-then-close operation.